Ajustement de l'outil de déplacement d'éléments Utilise dorénavant la commande de déplacement d'éléments.
